requirements.txt
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==3.0.5
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.0
daphne==3.0.2
decorator==5.1.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
executing==2.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
ipdb==0.13.13
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
msgpack==1.0.8
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
redis==5.0.7
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
wcwidth==0.2.13
zope.interface==6.4.post2
